refact: extrair window controls para WindowControlsListeners (actor-sheet.mjs)
